Copy feature (bug riddled)
I am too sleepy to fix this right now, but shift+d allows you to duplicate a tower. It is currently not working well with selection (I have no idea why) and for some reason it scales up the gizmo's sprites (the purple and red things with the area2D). Debug tomorrow.